+++ app/src/main/java/com/example/shiftproj/feature/repos/getTemp/IGetTemp.kt
@@ -2,5 +2,5 @@ package com.example.shiftproj.feature.repos.getTemp
 
 interface IGetTemp {
     fun isRealCity(city : String) : Boolean
-    fun getTemp(city : String) : Double
+    fun getTemp(city : String) : Long
 }
