+++ app/src/main/java/com/example/shiftproj/feature/repos/getTemp/TempRepos.kt
@@ -2,6 +2,6 @@ package com.example.shiftproj.feature.repos.getTemp
 
 object TempRepos : IGetTemp {
     private val dataStore : IGetTemp = TempGetterMock()
-    override fun getTemp(city: String): Double = dataStore.getTemp(city)
+    override fun getTemp(city: String): Long = dataStore.getTemp(city)
     override fun isRealCity(city: String): Boolean = dataStore.isRealCity(city)
 }
